app/pyproject.toml
[tool.poetry]
name = "brif"
version = "0.2.0"
description = "Docker based tool to deploy IIIF platforms."
authors = ["Pierre-Edouard Barrault"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"   # pinned after fastapi docker image
aiofiles = "==0.7.0"
celery = "==5.2.0"
iiif-prezi = "==0.3.0"
itsdangerous = "*"
Jinja2 = "==3.0.3"
numpy = "==1.21.4"
pydantic = "==1.8.2"
psycopg2-binary = "==2.9.2"
pytest = "==6.2.5"  # used by the test img
python-dotenv = "==0.19.2"
requests = "==2.26.0"
starlette-core = {git = "https://github.com/accent-starlette/starlette-core", branch = "master"}
typing = "==3.7.4.3"

[tool.poetry.dev-dependencies]
black = "==21.10b0"
flake8 = "==4.0.1"
isort = "==5.10.1"
pre-commit = "==2.15.0"
pylint = "==2.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
log_cli_level = "WARNING"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    # TODO: remove once Jinja is updated
    "ignore:'contextfunction' is renamed to 'pass_context', the old name will be removed in Jinja 3.1.:DeprecationWarning",
]
